recommender/requirements.txt
tensorflow>=2.15,<3.0
psycopg2-binary>=2.9
pgvector>=0.2
scikit-learn>=1.3
python-dotenv>=1.0
numpy>=1.24,<2.0
